Position int to decimal

+++ GameService.Application/Controller/RequestModel/PositionRequestModel.cs
@@ -4,9 +4,9 @@ namespace GameService.Controller.RequestModel
 {
     public class PositionRequestModel
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public decimal X { get; set; }
+        public decimal Y { get; set; }
+        public decimal Z { get; set; }
 
         public Position ToDomainModel()
         {
